Implement an enum for platforms
