+++ src/lib.rs
@@ -13,6 +13,7 @@ pub mod utils;
 pub use self::build_type::BuildType;
 pub use self::llvm_path::LLVMPath;
 pub use self::lock::Lock;
+pub use self::platforms::Platform;
 
 use std::path::PathBuf;
 use std::process::Command;
@@ -105,7 +106,7 @@ pub fn checkout(lock: Lock, force: bool) -> anyhow::Result<()> {
 ///
 pub fn build(
     build_type: BuildType,
-    targets: Vec<String>,
+    targets: Vec<Platform>,
     enable_tests: bool,
     enable_coverage: bool,
     extra_args: Vec<String>,
