+++ src/platforms/mod.rs
@@ -5,81 +5,48 @@
 pub mod aarch64_linux_gnu;
 pub mod aarch64_linux_musl;
 pub mod aarch64_macos;
+pub mod shared;
 pub mod x86_64_linux_gnu;
 pub mod x86_64_linux_musl;
 pub mod x86_64_macos;
 pub mod x86_64_windows_gnu;
 
-/// The build options shared by all platforms.
-pub const SHARED_BUILD_OPTS: [&str; 18] = [
-    "-DPACKAGE_VENDOR='Matter Labs'",
-    "-DCLANG_VENDOR='Matter Labs'",
-    "-DCLANG_REPOSITORY_STRING='origin'",
-    "-DCMAKE_COLOR_DIAGNOSTICS='Off'",
-    "-DLLVM_BUILD_DOCS='Off'",
-    "-DLLVM_INCLUDE_DOCS='Off'",
-    "-DLLVM_INCLUDE_BENCHMARKS='Off'",
-    "-DLLVM_INCLUDE_EXAMPLES='Off'",
-    "-DLLVM_ENABLE_DOXYGEN='Off'",
-    "-DLLVM_ENABLE_SPHINX='Off'",
-    "-DLLVM_ENABLE_OCAMLDOC='Off'",
-    "-DLLVM_ENABLE_ZLIB='Off'",
-    "-DLLVM_ENABLE_ZSTD='Off'",
-    "-DLLVM_ENABLE_LIBXML2='Off'",
-    "-DLLVM_ENABLE_BINDINGS='Off'",
-    "-DLLVM_ENABLE_TERMINFO='Off'",
-    "-DLLVM_ENABLE_LIBEDIT='Off'",
-    "-DLLVM_ENABLE_LIBPFM='Off'",
-];
+use std::str::FromStr;
 
-/// The build options shared by all platforms except MUSL.
-pub const SHARED_BUILD_OPTS_NOT_MUSL: [&str; 6] = [
-    "-DLLVM_DEFAULT_TARGET_TRIPLE='eravm'",
-    "-DLLVM_OPTIMIZED_TABLEGEN='On'",
-    "-DLLVM_BUILD_RUNTIME='Off'",
-    "-DLLVM_BUILD_RUNTIMES='Off'",
-    "-DLLVM_INCLUDE_RUNTIMES='Off'",
-    "-DLLVM_ENABLE_ASSERTIONS='On'",
-];
-
-/// The LLVM tests build options shared by all platforms.
-pub fn shared_build_opts_tests(enabled: bool) -> Vec<String> {
-    vec![
-        format!(
-            "-DLLVM_BUILD_UTILS='{}'",
-            if enabled { "On" } else { "Off" },
-        ),
-        format!(
-            "-DLLVM_BUILD_TESTS='{}'",
-            if enabled { "On" } else { "Off" },
-        ),
-        format!(
-            "-DLLVM_INCLUDE_UTILS='{}'",
-            if enabled { "On" } else { "Off" },
-        ),
-        format!(
-            "-DLLVM_INCLUDE_TESTS='{}'",
-            if enabled { "On" } else { "Off" },
-        ),
-    ]
+///
+/// The list of platforms used as constants.
+///
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum Platform {
+    /// The native X86 platform.
+    X86,
+    /// The native AArch64 platform.
+    AArch64,
+    /// The EraVM back end developed by Matter Labs.
+    EraVM,
+    /// The EVM back end developed by Matter Labs.
+    EVM,
 }
 
-/// The code coverage build options shared by all platforms.
-pub fn shared_build_opts_coverage(enabled: bool) -> Vec<String> {
-    vec![format!(
-        "-DLLVM_BUILD_INSTRUMENTED_COVERAGE='{}'",
-        if enabled { "On" } else { "Off" },
-    )]
+impl FromStr for Platform {
+    type Err = String;
+
+    fn from_str(value: &str) -> Result<Self, Self::Err> {
+        match value {
+            "EraVM" => Ok(Self::EraVM),
+            "EVM" => Ok(Self::EVM),
+            value => Err(format!("Unsupported platform: `{}`", value)),
+        }
+    }
 }
 
-/// Use of compiler cache (ccache) to speed up the build process
-pub fn shared_build_opts_ccache(use_ccache: bool) -> Vec<String> {
-    if use_ccache {
-        vec![
-            "-DCMAKE_C_COMPILER_LAUNCHER=ccache".to_string(),
-            "-DCMAKE_CXX_COMPILER_LAUNCHER=ccache".to_string(),
-        ]
-    } else {
-        vec![]
+impl std::fmt::Display for Platform {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            Self::X86 => write!(f, "X86"),
+            Self::AArch64 => write!(f, "AArch64"),
+            Self::EraVM => write!(f, "EraVM"),
+            Self::EVM => write!(f, "EVM"),
+        }
     }
 }
