+++ src/platforms/x86_64_linux_musl.rs
@@ -7,13 +7,14 @@ use std::process::Command;
 
 use crate::build_type::BuildType;
 use crate::llvm_path::LLVMPath;
+use crate::platforms::Platform;
 
 ///
 /// The building sequence.
 ///
 pub fn build(
     build_type: BuildType,
-    targets: Vec<String>,
+    targets: Vec<Platform>,
     enable_tests: bool,
     enable_coverage: bool,
     extra_args: Vec<String>,
@@ -183,13 +184,13 @@ fn build_musl(build_directory: &Path, target_directory: &Path) -> anyhow::Result
 /// The `crt` building sequence.
 ///
 fn build_crt(
-    mut targets: Vec<String>,
+    mut targets: Vec<Platform>,
     source_directory: &Path,
     build_directory: &Path,
     target_directory: &Path,
     use_ccache: bool,
 ) -> anyhow::Result<()> {
-    targets.push("X86".to_string());
+    targets.push(Platform::X86);
 
     crate::utils::command(
         Command::new("cmake")
@@ -209,7 +210,15 @@ fn build_crt(
                 "-DCMAKE_C_COMPILER='clang'",
                 "-DCMAKE_CXX_COMPILER='clang++'",
                 "-DLLVM_ENABLE_PROJECTS='compiler-rt'",
-                format!("-DLLVM_TARGETS_TO_BUILD='{}'", targets.join(";")).as_str(),
+                format!(
+                    "-DLLVM_TARGETS_TO_BUILD='{}'",
+                    targets
+                        .into_iter()
+                        .map(|platform| platform.to_string())
+                        .collect::<Vec<String>>()
+                        .join(";")
+                )
+                .as_str(),
                 "-DLLVM_DEFAULT_TARGET_TRIPLE='x86_64-pc-linux-musl'",
                 "-DLLVM_BUILD_TESTS='Off'",
                 "-DLLVM_BUILD_RUNTIMES='Off'",
@@ -227,8 +236,10 @@ fn build_crt(
                 "-DCOMPILER_RT_BUILD_MEMPROF='Off'",
                 "-DCOMPILER_RT_BUILD_ORC='Off'",
             ])
-            .args(crate::platforms::SHARED_BUILD_OPTS)
-            .args(crate::platforms::shared_build_opts_ccache(use_ccache)),
+            .args(crate::platforms::shared::SHARED_BUILD_OPTS)
+            .args(crate::platforms::shared::shared_build_opts_ccache(
+                use_ccache,
+            )),
         "CRT building cmake",
     )?;
 
@@ -310,8 +321,10 @@ fn build_host(
                 "-DCOMPILER_RT_DEFAULT_TARGET_ARCH='x86_64'",
                 "-DCOMPILER_RT_DEFAULT_TARGET_ONLY='On'",
             ])
-            .args(crate::platforms::SHARED_BUILD_OPTS)
-            .args(crate::platforms::shared_build_opts_ccache(use_ccache)),
+            .args(crate::platforms::shared::SHARED_BUILD_OPTS)
+            .args(crate::platforms::shared::shared_build_opts_ccache(
+                use_ccache,
+            )),
         "LLVM host building cmake",
     )?;
 
@@ -346,7 +359,7 @@ fn build_host(
 #[allow(clippy::too_many_arguments)]
 fn build_target(
     build_type: BuildType,
-    targets: Vec<String>,
+    targets: Vec<Platform>,
     source_directory: &Path,
     build_directory: &Path,
     target_directory: &Path,
@@ -388,7 +401,15 @@ fn build_target(
                 .as_str(),
                 "-DCMAKE_FIND_LIBRARY_SUFFIXES='.a'",
                 "-DCMAKE_EXE_LINKER_FLAGS='-fuse-ld=lld -static'",
-                format!("-DLLVM_TARGETS_TO_BUILD='{}'", targets.join(";")).as_str(),
+                format!(
+                    "-DLLVM_TARGETS_TO_BUILD='{}'",
+                    targets
+                        .into_iter()
+                        .map(|platform| platform.to_string())
+                        .collect::<Vec<String>>()
+                        .join(";")
+                )
+                .as_str(),
                 "-DLLVM_DEFAULT_TARGET_TRIPLE='eravm'",
                 "-DLLVM_OPTIMIZED_TABLEGEN='On'",
                 "-DLLVM_BUILD_RUNTIME='Off'",
@@ -397,13 +418,17 @@ fn build_target(
                 "-DLLVM_ENABLE_PROJECTS='llvm'",
                 "-DLLVM_ENABLE_ASSERTIONS='On'",
             ])
-            .args(crate::platforms::shared_build_opts_tests(enable_tests))
-            .args(crate::platforms::shared_build_opts_coverage(
+            .args(crate::platforms::shared::shared_build_opts_tests(
+                enable_tests,
+            ))
+            .args(crate::platforms::shared::shared_build_opts_coverage(
                 enable_coverage,
             ))
-            .args(crate::platforms::SHARED_BUILD_OPTS)
+            .args(crate::platforms::shared::SHARED_BUILD_OPTS)
             .args(extra_args)
-            .args(crate::platforms::shared_build_opts_ccache(use_ccache)),
+            .args(crate::platforms::shared::shared_build_opts_ccache(
+                use_ccache,
+            )),
         "LLVM target building cmake",
     )?;
